move cfn entities to separate file and group stack set ops in provider
